pyproject.toml
[tool.poetry]
name = "darf-core"
version = "0.1.0"
description = "Backend core platform for the Darf intelligence stack"
authors = ["darf"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.42"
psycopg2-binary = "^2.9.10"
langchain-community = "^0.3.27"
tiktoken = "^0.10.0"

[tool.poetry.group.dev.dependencies]
ruff = "0.3.7"
mypy = "^1.17.1"
pytest = "^8.4.1"
bandit = "^1.8.6"
safety = "^3.6.0"
pre-commit = "^4.2.0"
safety-cli = "^0.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = ["E", "F", "W", "B", "I", "C90", "UP"]
ignore = ["E501"]
target-version = "py312"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q"
testpaths = ["tests"]
